requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
ulid-py==1.1.0
urllib3==2.2.3
